[add] innerColor.jsx
